centralize auth logic in dedicated file
