+++ api/interface_adapters/controllers/repos.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 
 from api.domain.entities import Neighbour
 from api.infrastructure.external_services.github import GithubService
-from api.interface_adapters.gateways.user import get_current_active_user
+from api.interface_adapters.shared.auth import get_current_active_user
 from api.usecases.get_starneighbours import get_starneighbours as get_starneighbours_usecase
 from api.domain.errors import GithubRepositoryNotFound, GithubRateLimitExceeded
