+++ api/interface_adapters/gateways/user.py
@@ -1,15 +1,12 @@
 import os
-from fastapi import Depends, HTTPException, status
 from collections.abc import Iterator
 from sqlalchemy import Boolean, Column, Integer, String
 from sqlalchemy.exc import DatabaseError
 from sqlalchemy.orm import Session, sessionmaker
-from typing import Annotated
 
 from api.domain.entities import User
 from api.domain.repositories import IUserRepository
 from api.infrastructure.database.client import SQL_BASE, get_engine
-from api.interface_adapters.shared.auth import oauth2_scheme
 
 class UserInDB(SQL_BASE):
     __tablename__ = "app_user"
@@ -19,53 +16,6 @@ class UserInDB(SQL_BASE):
     hashed_password = Column(String(length=128), nullable=False)
     disabled = Column(Boolean, default=False)
 
-# TODO: seed database with those users and use
-fake_users_db = {
-    "johndoe": {
-        "id": 0,
-        "username": "johndoe",
-        "hashed_password": "fakehashedsecret",
-        "disabled": False,
-    },
-    "alice": {
-        "id": 1,
-        "username": "alice",
-        "hashed_password": "fakehashedsecret2",
-        "disabled": True,
-    },
-}
-
-def get_user(db, username: str):
-    if username in db:
-        user_dict = db[username]
-        return UserInDB(**user_dict)
-
-
-def fake_decode_token(token):
-    # This doesn't provide any security at all
-    # Check the next version
-    user = get_user(fake_users_db, token)
-    return user
-
-async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]):
-    user = fake_decode_token(token)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid authentication credentials",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    return user
-
-async def get_current_active_user(
-    current_user: Annotated[User, Depends(get_current_user)],
-):
-    if current_user.disabled:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
-
-# Implementation of IUserRepository interface
 class InMemoryUserRepository(IUserRepository):  # In-memory implementation of interface
     def __init__(self):
         self.data = {}
@@ -75,7 +25,6 @@ class InMemoryUserRepository(IUserRepository):  # In-memory implementation of in
 
     def get_by_username(self, username: str) -> User | None:
         return self.data.get(username)
-
 
 class SQLUserRepository(IUserRepository):  # SQL Implementation of interface
     def __init__(self, session):
